refactor: 가게 api prefix 수정
